requirements.txt
flask
openai
faiss-cpu
sentence-transformers
python-dotenv
numpy
scikit-learn
markdown
langchain
tiktoken
pytest
